# test: Cleanup Mul / Square benchmarking code