This also eliminates the 1 allocation per loop iteration in benchmarks
for `big.Int`.

```
name                     old time/op    new time/op    delta
Mul/single/uint256-6       18.0ns ± 1%    17.0ns ± 1%    -5.99%  (p=0.000 n=20+20)
Mul/single/big-6            211ns ± 0%     130ns ± 0%   -38.39%  (p=0.000 n=16+16)
Square/single/uint256-6    13.7ns ± 0%    12.9ns ± 1%    -5.77%  (p=0.000 n=18+20)
Square/single/big-6         204ns ± 0%     125ns ± 1%   -38.59%  (p=0.000 n=13+18)

name                     old alloc/op   new alloc/op   delta
Mul/single/uint256-6        0.00B          0.00B           ~     (all equal)
Mul/single/big-6            96.0B ± 0%      0.0B       -100.00%  (p=0.000 n=20+20)
Square/single/uint256-6     0.00B          0.00B           ~     (all equal)
Square/single/big-6         96.0B ± 0%      0.0B       -100.00%  (p=0.000 n=20+20)

name                     old allocs/op  new allocs/op  delta
Mul/single/uint256-6         0.00           0.00           ~     (all equal)
Mul/single/big-6             1.00 ± 0%      0.00       -100.00%  (p=0.000 n=20+20)
Square/single/uint256-6      0.00           0.00           ~     (all equal)
Square/single/big-6          1.00 ± 0%      0.00       -100.00%  (p=0.000 n=20+20)
```